pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tableau-agent-toolkit"
version = "0.1.0"
description = "Generate, validate, package, and publish Tableau workbooks from structured YAML specs"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "pydantic>=2.13",
    "typer>=0.25",
    "lxml>=6.1",
    "pyyaml>=6.0.3",
    "pydantic-settings>=2.14",
]

[project.scripts]
tableau-agent-toolkit = "tableau_agent_toolkit.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.1",
    "ruff>=0.15",
    "mypy>=2.0",
    "pre-commit>=4.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tableau_agent_toolkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
